Refactor: Add 'CreateEntities' module to enhance readability on 'app.rb'
- Add 'CreateEntities' module with the following methods:
  - create_person
  - create_student
  - create_teacher
  - create_book
  - create_rental

- Modified 'app.rb' to use the new module.
