+++ app.rb
@@ -1,9 +1,11 @@
-require_relative 'book'
-require_relative 'person'
-require_relative 'student'
-require_relative 'teacher'
-require_relative 'rental'
-require_relative 'classroom'
+require_relative "book"
+require_relative "person"
+require_relative "student"
+require_relative "teacher"
+require_relative "rental"
+require_relative "classroom"
+
+require_relative "modules/create_entities"
 
 class App
   attr_accessor :books, :people, :rentals
@@ -14,6 +16,29 @@ class App
     @rentals = []
   end
 
+  def create_person
+    person = CreateEntities.create_person
+    # * First it will promp the user to choose between student or teacher
+    # * Then it will return the person instance.
+    @people << person
+  end
+
+  def create_book
+    # * First it will promp the user for the creation of a book.
+    # * Then it will return the book instance.
+    book = CreateEntities.create_book
+    @books << book
+  end
+
+  def create_rental
+    rental = CreateEntities.create_rental(@books, @people)
+    # * First it will promp the user for the creation of a rental, by
+    # * iterating over the books and people arrays, and asking the user
+    # * to choose one of each.
+    # * Then it will return the rental instance.
+    @rentals << rental
+  end
+
   def list_books
     @books.each { |book| puts "Title: #{book.title}, Author: #{book.author}" }
   end
@@ -22,93 +47,16 @@ class App
     @people.each { |person| puts "#{person.class}: Name: #{person.name}, Id: #{person.id}, Age: #{person.age}" }
   end
 
-  def create_person
-    puts 'Do you want to create a student (1) or a teacher (2)? [Input the number]: '
-    person_type = gets.chomp
-
-    case person_type
-    when '1'
-      student = create_student
-      @people << student
-      puts 'Student created successfully'
-    when '2'
-      teacher = create_teacher
-      @people << teacher
-      puts 'Teacher created successfully'
-    else
-      puts 'Invalid option'
-    end
-  end
-
-  def create_student
-    print 'Name: '
-    name = gets.chomp
-
-    print 'Age: '
-    age = gets.chomp.to_i
-
-    print 'Classroom: '
-    classroom_name = gets.chomp
-    classroom = Classroom.new(classroom_name)
-
-    print 'Has parent permission? [Y/N]: '
-    parent_permission = gets.chomp.downcase == 'y'
-
-    Student.new(name, age, classroom, parent_permission: parent_permission)
-  end
-
-  def create_teacher
-    print 'Name: '
-    name = gets.chomp
-
-    print 'Age: '
-    age = gets.chomp.to_i
-
-    print 'Specialization: '
-    specialization = gets.chomp
-
-    Teacher.new(name, age, specialization)
-  end
-
-  def create_book
-    print 'Title: '
-    title = gets.chomp
-
-    print 'Author: '
-    author = gets.chomp
-
-    @books << Book.new(title, author)
-    puts 'Book created successfully'
-  end
-
-  def create_rental
-    puts 'Select a book from the following list by number:'
-    @books.each_with_index { |book, index| puts "#{index}) #{book.title}" }
-
-    book_index = gets.chomp.to_i
-
-    puts 'Select a person from the following list by number:'
-    @people.each_with_index { |person, index| puts "#{index}) #{person.name}" }
-
-    person_index = gets.chomp.to_i
-
-    print 'Date: '
-    date = gets.chomp
-
-    @rentals << Rental.new(date, @books[book_index], @people[person_index])
-    puts 'Rental created successfully'
-  end
-
   def list_rentals_for_person
-    print 'Enter person id: '
+    print "Enter person id: "
     person_id = gets.chomp.to_i
 
     rentals = @rentals.select { |rental| rental.person.id == person_id }
 
     if rentals.empty?
-      puts 'No rentals found for the given person id.'
+      puts "No rentals found for the given person id."
     else
-      puts 'Rentals:'
+      puts "Rentals:"
       rentals.each do |rental|
         puts "Date: #{rental.date}, Book: #{rental.book.title} by #{rental.book.author}"
       end
@@ -116,7 +64,7 @@ class App
   end
 
   def exit_app
-    puts 'Exiting the application. Goodbye!'
+    puts "Exiting the application. Goodbye!"
     exit
   end
 end
